app/ryan_cafe_app/services/orders_service/requirements.txt
Flask==3.1.1
mysql-connector-python==9.3.0
boto3==1.39.2
Werkzeug==3.1.3
gunicorn==23.0.0
